refac: 테스트를 역할에 따라 분리
